+++ changelog.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- No changes
+- Implemented method for setting enum properties to the device.
 
 ## [3.4.0] - 2017-10-17
 ### Added